Move env implementation to source file
