+++ CMakeLists.txt
@@ -50,6 +50,7 @@ set(iex_sources
         iex.cc
         curl_wrapper.cc
         file_serializable.cc
+        env.cc
         )
 
 # Specify source directory.
